+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [UNRELEASED]
+
+- correctly rendering the context around the match line in grep power tool output
+
 ## [0.66.0]
 
 - added limit truncation to fetch tool result